added tool rtl_raw2wav
* rtl_raw2wav: similar to raw 2 wav conversion with sox,
  but additional chunk with frequency/time information
* splitted convenience.h/.c, added rtl_convenience.h/.c
