+++ src/convenience/wavewrite.c
